refactor!: move error codes to @libp2p/interface/errors
