+++ packages/libp2p/src/config.ts
@@ -1,9 +1,9 @@
-import { CodeError } from '@libp2p/interface/errors'
+import { CodeError, codes } from '@libp2p/interface/errors'
 import { FaultTolerance } from '@libp2p/interface/transport'
 import { defaultAddressSort } from '@libp2p/utils/address-sort'
 import { dnsaddrResolver } from '@multiformats/multiaddr/resolvers'
 import mergeOptions from 'merge-options'
-import { codes, messages } from './errors.js'
+import { messages } from './errors.js'
 import type { Libp2pInit } from './index.js'
 import type { ServiceMap, RecursivePartial } from '@libp2p/interface'
 import type { Multiaddr } from '@multiformats/multiaddr'
