+++ packages/libp2p/src/transport-manager.ts
@@ -1,8 +1,7 @@
-import { CodeError } from '@libp2p/interface/errors'
+import { CodeError, codes } from '@libp2p/interface/errors'
 import { trackedMap } from '@libp2p/interface/metrics/tracked-map'
 import { FaultTolerance } from '@libp2p/interface/transport'
 import { logger } from '@libp2p/logger'
-import { codes } from './errors.js'
 import type { Libp2pEvents, AbortOptions } from '@libp2p/interface'
 import type { Connection } from '@libp2p/interface/connection'
 import type { EventEmitter } from '@libp2p/interface/events'
